Refactor the gds/hash code for ease of editing
